feat: redesign codec header layout and self-describing Unspecified encoding
  Rebalance the 8-byte DataHeader from 4×u16 to count(u32) + blob_size(u16) +
  data_fields(u8) + ordinal(u8), giving count the range it needs while right-sizing
  ordinal and data_fields. System ordinals now count down from 255; user ordinals
  start at 1. Replace DynamicDataValue/DynamicListValue/DynamicMapValue with a
  self-describing Unspecified codec that round-trips any typed data opaquely.

+++ codas/src/parse.rs
@@ -34,8 +34,9 @@ pub fn parse(markdown: &str) -> Result<Coda, ParseError> {
 
     // Create data types.
     for (ordinal, parsed_data) in parsed_coda.data.into_iter().enumerate() {
-        // Ordinals are 1-indexed.
-        let ordinal = (ordinal + 1) as u16;
+        // User-defined ordinals start at 1 (0 reserved for Unspecified,
+        // 224-255 reserved for built-in system types).
+        let ordinal = (ordinal + 1) as u8;
 
         // Extract docs.
         let docs = if parsed_data.docs.is_empty() {
